Implement Sentence class
Specification
In the Sentence class, complete the implementations of known_mines, known_safes, mark_mine, and mark_safe.

The known_mines function should return a set of all of the cells in self.cells that are known to be mines.
The known_safes function should return a set of all the cells in self.cells that are known to be safe.
The mark_mine function should first check to see if cell is one of the cells included in the sentence.
If cell is in the sentence, the function should update the sentence so that cell is no longer in the sentence, but still represents a logically correct sentence given that cell is known to be a mine.
If cell is not in the sentence, then no action is necessary.
The mark_safe function should first check to see if cell is one of the cells included in the sentence.
If cell is in the sentence, the function should update the sentence so that cell is no longer in the sentence, but still represents a logically correct sentence given that cell is known to be safe.
If cell is not in the sentence, then no action is necessary.