Implement make_safe_move and make_random_move in MinesweeperAI
Specifications

make_safe_move should return a move (i, j) that is known to be safe.
The move returned must be known to be safe, and not a move already made.
If no safe move can be guaranteed, the function should return None.
The function should not modify self.moves_made, self.mines, self.safes, or self.knowledge.

make_random_move should return a random move (i, j).
This function will be called if a safe move is not possible: if the AI doesn’t know where to move, it will choose to move randomly instead.
The move must not be a move that has already been made.
The move must not be a move that is known to be a mine.
If no such moves are possible, the function should return None.

NOTE : make_random_move is rudimentary. It would look for a random move for 100 times before returning None. There is no guarantee that a possible move is not found after 100 attemps, but it is a simple way to avoid an endless loop...